# CMTChattingServer - define의 전처리 constexpr 로 바꾸기 - 타입변환시 크기 안맞던 경고 해결
#CNetworkLib , CMTChattingServer , clsSession
- 템플릿 암시적 생성 방지 문구를 stTlsObjectPool 헤더에서 명시할수 있는 방법을 찾았음.
>>>>>>>>>>>> 템플릿 특수화에 관련하여 공부하기.

#각 생성자들의 초기화 경고 제거. 안쓰는 지역변수 제거

+++ _Portfolio/_ChattingServerProject/MT_ChattingServer/CNetworkLib/CNetworkLib.h
@@ -70,7 +70,7 @@ class CLanServer : public Stub, public Proxy
     void SessionUnLock(ull SessionID); // 반환형 쓸때가 없음.
 
     void SendPacket(ull SessionID, struct CMessage *msg, BYTE SendType,
-                    std::vector<ull> *pIDVector = nullptr, WORD wVecLen = 0);
+                    std::vector<ull> *pIDVector = nullptr, size_t wVecLen = 0);
     void Unicast(ull SessionID, CMessage *msg, LONG64 Account = 0);
     void BroadCast(ull SessionID, CMessage *msg, std::vector<ull> *pIDVector, WORD wVecLen);
